+++ lib/Controls/ControlTypes/Triggers/Trigger.js
@@ -377,6 +377,8 @@ export default class ControlTrigger extends ControlBase {
 		const allActions = this.actions.getAllActions()
 
 		this.internalModule.findReferencedInstances(foundInstanceIds, foundInstanceLabels, allActions, allFeedbacks)
+
+		// TODO - ideally this should consider references from module actions/feedbacks.
 	}
 
 	/**
