+++ lib/Controls/Fragments/FragmentActions.js
@@ -435,12 +435,12 @@ export default class FragmentActions extends CoreBase {
 
 	/**
 	 * Get the instance ids the actions belong to
-	 * @param {Set<String> | undefined} targetSet Optional set to insert into
+	 * @param {Set<String> | undefined} targetInstanceIdsSet Optional set to insert into
 	 * @access public
 	 * @returns {Set<string>}
 	 */
-	getReferencedInstances(targetSet) {
-		const instanceIds = targetSet || new Set()
+	getReferencedInstances(targetInstanceIdsSet) {
+		const instanceIds = targetInstanceIdsSet || new Set()
 
 		for (const action_set of Object.values(this.action_sets)) {
 			for (const action of action_set) {
