+++ lib/Controls/Fragments/FragmentFeedbacks.js
@@ -466,12 +466,12 @@ export default class FragmentFeedbacks extends CoreBase {
 
 	/**
 	 * Get the instance ids the feebacks belong to
-	 * @param {Set<String> | undefined} targetSet Optional set to insert into
+	 * @param {Set<String> | undefined} targetInstanceIdsSet Optional set to insert into
 	 * @access public
 	 * @returns {Set<string>}
 	 */
-	getReferencedInstances(targetSet) {
-		const instanceIds = targetSet || new Set()
+	getReferencedInstances(targetInstanceIdsSet) {
+		const instanceIds = targetInstanceIdsSet || new Set()
 
 		for (const feedback of this.feedbacks) {
 			instanceIds.add(feedback.instance_id)
