+++ lib/Data/ImportExport.js
@@ -735,9 +735,13 @@ class DataImportExport extends CoreBase {
 		// Future: this does not feel durable
 
 		const instanceLabelRemap = {}
-		for (const info of Object.values(instanceIdMap)) {
+		const instanceIdRemap = {}
+		for (const [oldId, info] of Object.entries(instanceIdMap)) {
 			if (info.label !== info.oldLabel) {
 				instanceLabelRemap[info.oldLabel] = info.label
+			}
+			if (info.id && info.id !== oldId) {
+				instanceIdRemap[oldId] = info.id
 			}
 		}
