+++ lib/Internal/Controller.js
@@ -28,8 +28,8 @@ import Variables from './Variables.js'
 import { cloneDeep } from 'lodash-es'
 import { ParseControlId } from '../Shared/ControlId.js'
 import Page from './Page.js'
-import { SplitVariableId } from '../Resources/Util.js'
-import { findVariableInstancesInString, replaceVariableInstancesInString } from '../Instance/Variable.js'
+import { InternalReferencesUpdater } from './Visitors/ReferencesUpdater.js'
+import { InternalReferencesCollector } from './Visitors/ReferencesCollector.js'
 
 export default class InternalController extends CoreBase {
 	feedbacks = new Map()
@@ -328,89 +328,6 @@ export default class InternalController extends CoreBase {
 			if (typeof fragment.variablesChanged === 'function') {
 				fragment.variablesChanged(changed_variables, removed_variables)
 			}
-		}
-	}
-}
-
-export class InternalReferencesCollector {
-	constructor(foundInstanceIds, foundInstanceLabels) {
-		this.instanceLabels = foundInstanceLabels || new Set()
-		this.instanceIds = foundInstanceIds || new Set()
-	}
-
-	applyExpression(obj, propName) {
-		this.applyString(obj, propName)
-	}
-
-	applyInstanceId(obj, propName) {
-		this.instanceIds.add(obj[propName])
-	}
-	applyInstanceIdArray(obj, propName) {
-		for (const id of obj[propName]) {
-			foundInstanceIds.add(id)
-		}
-	}
-
-	applyString(obj, propName) {
-		const labels = findVariableInstancesInString(obj[propName])
-		for (const label of labels) {
-			foundInstanceLabels.add(SplitVariableId(label)[0])
-		}
-	}
-
-	applyVariableName(obj, propName) {
-		const label = SplitVariableId(obj[propName])[0]
-		this.instanceLabels.add(label)
-	}
-}
-export class InternalReferencesUpdater {
-	constructor(instanceLabelsRemap, instanceIdRemap) {
-		this.instanceLabelsRemap = instanceLabelsRemap
-		this.instanceIdRemap = instanceIdRemap
-		this.changed = false
-	}
-
-	applyExpression(obj, propName) {
-		this.applyString(obj, propName)
-	}
-
-	applyInstanceId(obj, propName) {
-		if (!this.instanceIdRemap) return
-
-		const oldId = obj[propName]
-		const newId = this.instanceIdRemap[oldId]
-		if (newId && newId !== oldId) {
-			obj[propName] = newId
-			this.changed = true
-		}
-	}
-	applyInstanceIdArray(obj, propName) {
-		if (!this.instanceIdRemap) return
-
-		const array = obj[propName]
-		for (let i = 0; i < array.length; i++) {
-			this.applyInstanceId(array, i)
-		}
-	}
-
-	applyString(obj, propName) {
-		if (!this.instanceLabelsRemap) return
-
-		const rawStr = obj[propName]
-		if (typeof rawStr !== 'string') return
-
-		obj[propName] = replaceVariableInstancesInString(this.instanceLabelsRemap, rawStr)
-	}
-
-	applyVariableName(obj, propName) {
-		if (!this.instanceLabelsRemap) return
-
-		const id = SplitVariableId(obj[propName])
-		const newLabel = this.instanceLabelsRemap[id[0]]
-		console.log('remap', newLabel, obj, propName)
-		if (newLabel) {
-			obj[propName] = `${newLabel}:${id[1]}`
-			this.changed = true
 		}
 	}
 }
