+++ lib/Internal/Controls.js
@@ -22,8 +22,6 @@ import { SplitVariableId, rgb } from '../Resources/Util.js'
 import { CreateBankControlId } from '../Shared/ControlId.js'
 import { ButtonStyleProperties } from '../Shared/Style.js'
 import debounceFn from 'debounce-fn'
-import { FindAllReferencedVariablesInExpression } from '../Shared/Expression/ExpressionParse.js'
-import { findVariableInstancesInString } from '../Instance/Variable.js'
 
 const previewControlIdFn = ((options, info) => {
 	// Note: this is manual, but it can't depend on other functions
