+++ lib/Shared/Expression/ExpressionParse.js
@@ -83,13 +83,3 @@ export function FindAllReferencedVariables(node) {
 
 	return referencedVariables
 }
-
-/**
- * Find any variables referenced in an expression
- * @param {string} expression
- * @returns {Array<string>}
- */
-export function FindAllReferencedVariablesInExpression(expression) {
-	const node = ParseExpression(expression)
-	return FindAllReferencedVariables(node)
-}
